refactor: remove redundant files
