--- liquidator/src/amm/PoolDataProviderInterface.ts
@@ -1,7 +1,0 @@
-import { Pool } from "../types/Pool";
-
-export interface PoolDataProviderInterface {
-
-    loadPools(): Promise<Pool[]>
-
-}
